添加检查Cookie是否有效, 添加某些错误提示!
